Extract `TabBar` into a reusable component, outside of application logic.
